Rewrite interface to use WindowFunction instead of FlatMapFunction
Previously, it was necessary to collect all elements which should be aggregated via making a set of a single element and than combining those via reduction to one set of all to-be-aggregated elements.
This change elliminates this process by using WidowFunction instead, which provides an iterator over all elements in the window.